Changed the implementation of capital costs of new plants to make it easier to determine the payments due in any given year.
Tweaked implementation of capital payments for existing plants to slightly simplify the model.

Changed implementation of fixed O&M for new plants to be indexed in the same way as existing to be able to describe the objective function in fewer lines.

Made all CRF equations have the same form so they will be easier to read and debug.

Fixed an error in the calculation of capital costs of keeping existing plants online past their retirement. 

Added a "ConsumePower" decision variable that forces the decision of which type of power to consume and which to spill to fully satisfy RPS goals. (RPS goals are defined by power consumed, not power available)

Added a number of derived variables to simplify the model and avoid recalculating the same value in many different places. These variables are factored out by AMPL during the compilation stage (before it sends the problem to cplex). They add to the complexity of the compilation time, don't add to the complexity of the resultant optimization program, and reduce the complexity of the code. The new variables are: Installed_To_Date, Installed_To_Date_by_Period, Power_Produced, EP_Power_Produced, Gross_Power, and Net_Power. 
Updated the objective function, constraints, and reporting code to take advantage of the new variables.

Added a constraint "Maximum_Store_Pumped_Hydro_Reserve" that was previously neglected.

Added a number of reporting variables that take over calculations that were performed in basicstats and record_results to make the code easier to read as well as permiting easier inspection of results on the AMPL command line. These variables are defined as "parameters" in windsun.run. After a solution is calculated or loaded from disk, their values are set in update_reporting_variables.run.

I changed cplex options to report the results of presolve simplifications and I explicitly the presolve option to its default value of 1 to clarify program behavior. 

I added AMPL options "linelim 1" and "substout 1" to have AMPLE substitute out the derived decision variables before passing the problem to CPLEX while maintaining linearity. 

Added functionality to track and report how long AMPL takes to compile the linear program and write it to disk. This helps with performance profiling. 

Added logic to make the results directory if it doesn't exist. This enabled me to remove the empty results directory from subversion. 

Made use of AMPL's "named problem" feature to handle the distinction between minimizing cost and transmission usage. 

Removed extraneous parameters that have not been used in years.

Moved the functionality that provides parameter checks for the Min_Solar_Production constraint from windsun.run and switch.run into a new file min_solar_production_check.run.

Jimmy's DatabasePrep changes: Updated cogen heat rates to reflect values from the EIA database. Previously they were inferior assumptions from Matthias's time. Also renamed a few columns for clarity.