Corrected financial calculations that converted a future series of payments to NPV. The Capital Recovery Factor (CRF) calculates the value in the year before the first payment. We had been mistakenly treating values based on the CRF as if they were in the year of the first payment.
Corrected variable costs. We had been multiplying the variable costs incurred in the first year of the period by the number of years per period. We have been applying discounting within a period to all other costs (fixed O&M, capital). This systematically overestimated the variable costs relative to the other costs. The upshot is that this lowered variable costs by 10-20 % (depending on the length of the period and using our standard 7% discount rate). We also applied this correction to variable carbon costs (carbon_cost_per_mwh and ep_carbon_cost_per_mwh).