I modified shell calls to be in the form of functions to make it easier to retrieve results and obviate the need for temporary files which were causing race conditions on the cluster. These functions are get_num_cores() and file_exists(). I explain the rational for their peculiar implementation in AMPL_errata.txt under the new directory 'doc'.
Added better job management for carbon curves in a cluster context. There are now two function that will return the number of workers and give individual workers their id. (This functionality is robust: when ran outside of a cluster environment, the functions report a sole worker.) If two workers are tasked with a problem, then one will work on every other carbon cost and the other will do the remainder.
I also removed the transmission optimization code from switch.run. We can retrieve it from subversion later, and in the meanwhile I was too lazy to propogate these changes to commented-out code.