Changed code to optionally allow AMPL to write the problem file and exit, then have cplex solve the optimization, then call AMPL again to parse cplex's solution. Calling cplex from the command line instead of from AMPL can save significant amounts of memory. To use this funcitonality, invoke AMPL with run_switch.sh
run_switch.sh now will work on or off of the cluster. Off of the cluster, it assumes that you are starting AMPL in single-thread mode.

There is a new parameter in windsun.run compile_mip_only that defaults to 0. If it is set to 1, then switch.run won't invoke cplex with the solve command, but will instead compile problems, parse any solution files it finds, and write cplex options out to a file.