This adds spinning and quickstart operating reserves to the model.
The balancing areas are assumed to be the NERC primary subregions. The six balancing subregions are: California, Pacific Northwest, Arizona-New Mexico-Nevada, Rocky Mountains, Canada, and Baja Mexico.

The amount of spinning reserves required is based on the 5-3 rule from the Western Wind and Solar Integration Study. It is assumed to be the sum of 3 percent of load, 5 percent of wind generation, and 5 percent of non-CSP solar generation in each hour. The amount of quickstart capacity required is assumed to be the same as the spinning reserve requirement plus 5 percent of generation from CSP with no storage. CSP with storage is assumed to have enough thermal inertia not to contribute to either the spinning or quickstart reserve requirement.

Thermal generators, including gas steam turbines, gas combustion turbines, and CCGT as well as distillate fuel oil turbines and engines, are assumed to be able to provide spinning reserves at a heat rate penalty and quickstart reserves at no additional cost.  The heat rate penalty calculation is in the generator_costs.xlsx spreadsheet and is based on generator part load efficiencies at 75 percent of full load. Thermal generators can only provide spinning reserves if they are also dispatching useful energy in that hour. The amount of spinning reserves that can be provided is limited to less than the generator 10-min ramp rates in all cases. In particular, 25 percent was used as the maximum capacity available for spinning reserve from steam turbines and CCGT, and 30 percent for spinning reserve from combustion turbines (the actual 10-min ramp rates are higher, but the heat rate penalty incurred by backing off useful generation further would be prohibitive).

Hydro projects, pumped hydro storage projects, and storage projects are also assumed to be able to provide operating reserve. Since these types of generators have very fast response rates, all of their operating reserve is assumed to be able to count toward the spinning reserve requirement. Once the spinning reserve requirement is met, any additional operating reserve provided by hydro, pumped hydro, and storage can be used to meet the quickstart reserve requirement.

Hydro operating reserve is limited to no more than 20 percent of the project’s capacity.

It is assumed that operating reserves from hydro and storage will be dispatched 1 percent of the time, and this is included in the projects’ respective energy balance constraints.

In reporting the results, the operating reserve provided by each hydro, pumped hydro, and storage project is apportioned between a spinning and a quickstart component based on the ratio for the whole balancing area (which is calculated by subtracting the thermal spinning reserve provided from the total spinning reserve requirement in the balancing area in each hour to figure out how much of the generic hydro/pumped hydro/storage operating reserve was used as spinning). For CAES, the amount of storage operating reserve that was spinning or quickstart is based on the spinning and quickstart reserve provided by the natural gas component of the storage plant via the NG to storage ratio.

The total amount of useful generation, spinning reserve, and quickstart capacity cannot exceed the generator size, derated for forced outages.