Implemented EV demand response capability.
It is similar to the modeling of DR from residential and commercial heating, but the two are kept separate. The EV demand response scenarios are in the ev_scenarios table and are linked to the load_scenario_ids (combinations must be unique). An ev_scenario_id has been added to the scenarios_v3 table. Old scenarios have an ev_scenario_id value of NULL and will not be affected as the new code should be backward-compatible. If the ev_scenario_id is NULL, no input tab file is exported from the database, the EV DR variables are fixed, and the associated constraints are dropped.