Changed the way energy balance is implemented for new storage projects from daily balancing to timepoint-to-timepoint tracking (still within the day with the last hour of the day assumed to be the previous timepoint for the first hour of the day). The total amount of energy in storage at any point is limited to InstallGen * hours of storage duration (15 hours for CAES and 8.1 hours for batteries). It is assumed that in order to provide operating reserves, a storage project must be able to sustain that level of output for an hour, so that amount of energy must be available in the timepoint when reserves are provided.
Also, removed the accounting for the fraction of the time that operating reserves are dispatched from the storage and hydro energy constraints. The reasoning is that, when dispatched, the operating reserves are simply replacing generation that has gone down, so the net extra generation needed to meet load is zero. Since we don't actually model outages, the previous implementation resulted in the dispatch of extra generation.

We still need to make sure that there is sufficient energy available from hydro and pumped storage in every timepoint to be able to provide reserves given the duration requirement for reserves. We'll do that in a separate commit.