Implemented a two-variable treatment of new storage: decisions are now made for both the power capacity and the energy capacity (duration) of storage projects. Each has associated costs, so the total cost of a storage project is now calculated as the installed power capacity times a per MW cost plus the installed energy capacity times a per MWh cost. The duration of storage can then be calculated as the installed energy capacity divided by the installed power capacity.
Costs are based on B&V for batteries and on EPRI for CAES (see the storage_params sheet in the generator_info_and_costs spreadhsheet for details).

I did not yet propagate to dispatch the ability to increase the energy capacity of storage in case of infeasibilities (power capacity can still be increased).