Implemented the ability to optimally dispatch CSP with storage in SWITCH.
The prior implementation was an exogenously-specified dispatch pattern.

The new implementation starts with the data on energy availability for CSP without storage, which is assumed to have a solar multiple of 1.4. Systems with 6 hours of storage are assumed to have a solar multiple of 2. To obtain the hourly energy availability for projects with 6 hours of storage, these projects were matched via their location_id to projects with no storage, and were assigned hourly energy availability from the solar field of [ ( 2/1.4 ) * the solar field energy availability of the projects without storage ]. WHY ONLY 2006?

Due to computational constraints, the hourly dispatch of CSP with storage is aggregated to the load area level. It is then apportioned back to the project level in basicstats.run based on relative capacity. The variables optimal CSP dispatch are DispatchCSP, StoreCSP, ReleaseCSP, and TotalEnergyinCSPStorage. DispatchCSP is the energy that is dispatched directly to the turbine from the solar field without being stored first.

Allowing parasitic losses greatly increased runtime, so parasitic losses from CSP plants with storage are ignored in this implementation. As hourly cap factors are negative in some hours in the input data, ignoring parasitic losses is accomplished by defining DispatchCSP and StoreCSP as >= 0 and constraining the sum of DispatchCSP and StoreCSP (i.e. the total energy available from the solar fields) to be <= the max of ( the hourly cap factor * installed capacity ) and 0.

Power output from CSP with storage (the sum of DispatchCSP and ReleaseCSP) is limited by turbine size.

The TotalEnergyinCSPStorage can never exceed 6 hours of storage (6 h * MW capacity). The energy in storage in each hour is tracked in the same was as it is for batteries and CAES.

The variable cost of CSP with storage is taken to be /MWh based on Hummon et al, 2013, Modeling CSP with TES for Integration Studies, NREL/CP-6A20-60365. I picked a value in the middle between the low and high estimates. Variable cost is assumed to be incurred upon dispatching energy directly to the turibne (DispatchCSP) and upon storing energy (StoreCSP), the latter in order to avoid spilling energy in the form of inefficient CSP storage dispatch. Like with batteries and CAES, the variable cost is assumed to be per MWh-output, so StoreCSP is multiplied by storage efficiency as that's how much energy will be released.

An additional change on the database side is that CSP_Trough_6h_Storage now has a value defined for storage_efficiency in generator_info.

Currently, only the results for hourly power output of CSP with storage (the sum of DispatchCSP and ReleaseCSP) are imported into the results database. StoreCSP is not imported. This is OK as an important distinction between StoreCSP and StoreEnergy/StorePumpedHydro is that the energy stored in CSP thermal storage is not taken from the grid.