Implemented hourly, load-area-level Spill and Consume variables. This makes it possible to more easily track how much energy is spilled after the transmission optimization is completed.
The variables are implemented separately for distributed and non-distributed power
because of the difference in treatment of distribution losses for the two.

The sum of Consume_NonDistributed and Consume_Distributed must be equal to system load.

The sum of ConsumeDistributed and SpillDistributed must equal in-load-area
distributed generation in each hour.

The sum of ConsumeNonDistributed and SpillNonDistributed must equal the in-load-area
non-distributed generation plus net transmission in each hour.

The reserve consume variables were renamed Contribute_To_Planning_Reserve for clarity.

These are made equal to the amount of planning reserves provided by in-load-area
distributed and non-distributed projects respectively,
plus net transmission for the non-distributed.

The sum of the new spill variables is used in basicstats.run instead of the old method
for calculating spilled energy. This may still be an understimate as transmission has not yet been optimized when basicstats is run.

Actual spilled energy is calculated after the transmission optimization is complete
(the spill and consume variables are not fixed in the transmission optimization)
and then imported into the database.